+++ activity/brightpay/react/changelog.md
@@ -4004,3 +4004,14 @@
 
 ---
 
+# Brightpay Development - 2025-10-15
+
+**Commit #365**
+
+- Repository: brightpay/checkout
+- Type: payment-system
+- Language: react
+- Month: 10/2025
+
+---
+